+++ src/app/pages/project-navigator/project-navigator.component.ts
@@ -1,6 +1,9 @@
 import { Component } from '@angular/core';
 import { NavigatorFilterSelectorModule,NavigatorObjectiveStatusFilterModule,NavigatorRangeFilterModule,NavigatorUsersFilterModule } from '@indziaki/ngx-xchange-ui';
-
+export interface Collaborator {
+  id: number;
+  name: string;
+}
 @Component({
   selector: 'app-project-navigator',
   standalone: true,
@@ -14,7 +17,7 @@ import { NavigatorFilterSelectorModule,NavigatorObjectiveStatusFilterModule,Navi
   styleUrl: './project-navigator.component.scss'
 })
 export class ProjectNavigatorComponent {
-  collaborator: any[] = [
+  collaborator: Collaborator[] = [
     { id: 1, name: 'User 1' },
     { id: 2, name: 'User 2' },
     { id: 3, name: 'User 3' }